+++ src/lasuite/oidc_login/enums.py
@@ -1,0 +1,12 @@
+"""Enum for OIDC login."""
+
+from enum import StrEnum
+
+
+# OIDC_OP_USER_ENDPOINT allowed values
+class OIDCUserEndpointFormat(StrEnum):
+    """Enum for OIDC OP User Endpoint."""
+
+    JWT = "jwt"
+    JSON = "json"
+    AUTO = "auto"