+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- ✨(oidc) allow JSON format in user info endpoint #5
+
 ## [0.0.2] - 2025-04-07
 
 ### Fixed